% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_bariatricRat.binned.5.R
\docType{data}
\name{bariatricRat.binned.5}
\alias{bariatricRat.binned.5}
\title{NMR spectra of urine sample of rats with bariatric surgery. A complete description of the dataset can be found here: https://www.frontiersin.org/articles/10.3389/fmicb.2011.00183}
\format{
A data frame with 59 rows, X, metadata and ppm scale
\describe{
\item{data}{NMR spectra}
\item{binned.ppm}{a ppm scale}
}
}
\source{
{https://www.frontiersin.org/articles/10.3389/fmicb.2011.00183}
}
\usage{
bariatricRat.binned.5
}
\description{
The data set contains 59 nmr spectra of length 4582 It includes metadata (425 exprimental parameters)
}
\keyword{datasets}
