% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_bariatricRat.R
\docType{data}
\name{bariatricRat}
\alias{bariatricRat}
\title{NMR spectra of urine sample of rats with bariatric surgery. A complete description of the dataset can be found here: https://www.frontiersin.org/articles/10.3389/fmicb.2011.00183}
\format{
A data frame with 59 rows, X, metadata and ppm scale
\describe{
\item{metadata}{metadata for each sample}
\item{X}{NMR spectra}
\item{ppm}{a ppm scale}
}
}
\source{
{https://www.frontiersin.org/articles/10.3389/fmicb.2011.00183}
}
\usage{
bariatricRat
}
\description{
The data set contains 59 nmr spectra of length 22913. It includes metadata (425 exprimental parameters)
}
\keyword{datasets}
