% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMetadata.R
\name{checkMetadata}
\alias{checkMetadata}
\title{checkMetadata}
\usage{
checkMetadata(dataFrame, option = list())
}
\arguments{
\item{dataFrame}{\itemize{
\item metadata
}}

\item{option}{\itemize{
\item options
}}
}
\value{
a data.frame object
}
\description{
This function check and prepare a metadata dataframe to be used within hastaLaVista
framework. It may contain as much column as necessary, but should be of type
data.frame and MUST contain at least one column with unique values. This column
should be named "sampleID". Thus each row of the metadata object describe one sample
}
\examples{
N <- 500
x <- rnorm(N, mean=0, sd=10)
y <- rnorm(N, mean=0, sd=10)
sampleID <- paste0("sampleID", seq_along(x))
sampleGroup <- paste0("sampleGroup", seq_along(x))
sampleColor <- c(rep(1,N/2), rep(2,N/2))
metadata <- data.frame(sampleID, sampleGroup, sampleColor)
checkMetadata(metadata)
}
