% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_coffeeSpectra.R
\docType{data}
\name{coffeeNMRSpectra}
\alias{coffeeNMRSpectra}
\title{NMR spectra of coffee beans from Colombia, Peru and Brazil}
\format{A data frame with 34 rows and 4 variables:
\describe{
  \item{param}{metadata for each sample}
  \item{nmr}{NMR spectra}
}}
\source{
{Comparison of Attenuated Total Reflectance Mid-Infrared, Near Infrared, and 1H-Nuclear Magnetic Resonance Spectroscopies for the Determination of Coffee’s Geographical Origin}
\url{https://doi.org/10.1155/2017/7210463 }
}
\usage{
coffeeNMRSpectra
}
\description{
The data set contains nmr spectra of 34 sample from 3 countries.
This dataset is a subset of one published \url{https://doi.org/10.1155/2017/7210463 } and the experimental section can be taken as a reference.
}
\keyword{datasets}
