% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.NMRium.R
\name{exportReIm}
\alias{exportReIm}
\title{Method for exporting spectra information into the NMRium format. You can use this to export simple x, re and x, re and im spectra
If you provide a list containing NMRium data, it will append the new spectra data into the proper place,
so that you can get an overlay of several spectra into the platform.}
\usage{
exportReIm(
  x,
  yr,
  yi,
  output = NA,
  observeFrequency = 600,
  dataType = "NMR SPECTRUM",
  solvent = "UNKNOWN",
  nucleus = "1H",
  PHC0 = 0,
  PHC1 = 0,
  col = "#000000"
)
}
\arguments{
\item{x}{Independent variable}

\item{yr}{Real dependent variable}

\item{yi}{(optional)Imaginary dependent variable}

\item{output}{Object containing other NMRium data}

\item{observeFrequency}{Observe frequency in MHz}

\item{dataType}{Data type format. Use 'NMR FID' if the spectrum is in the time domain (Default 'NMR SPECTRUM')}

\item{solvent}{String that specify the solvent. (Default 'UNKNOWN')}

\item{nucleus}{Observe nucleus.  (Default '1H')}

\item{PHC0}{Phase 0 correction value (Optional, Default 0)}

\item{PHC1}{Phase 1 correction value (Optional, Default 0)}

\item{col}{Color for this spectrum(Default black)}
}
\value{
list
}
\description{
Method for exporting spectra information into the NMRium format. You can use this to export simple x, re and x, re and im spectra
If you provide a list containing NMRium data, it will append the new spectra data into the proper place,
so that you can get an overlay of several spectra into the platform.
}
