% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColor2.R
\name{getColor2}
\alias{getColor2}
\title{GetColorChart based on character}
\usage{
getColor2(text, colorFormat = "rgb")
}
\arguments{
\item{text}{a text used to pick the color}

\item{colorFormat}{the format color should be output. Current options are rgb and hex.}
}
\value{
a rgba color string
}
\description{
GetColorChart based on character
}
\examples{
getColor2("colombia", "hex")
}
