% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initServer.R
\name{initServer}
\alias{initServer}
\title{checks if server is running and if not starts it}
\usage{
initServer(server, force = FALSE)
}
\arguments{
\item{server}{a server object}

\item{force}{if true, will force a second server on a new port. Should not be necessary in regular use.}
}
\value{
void
}
\description{
This functions starts the http server that is necessary to the \code{github.com/npellet/visualizer} to work.
If a server is already running (daemon_list is empty), then nothing happens.
}
\examples{
#start new server object
s <- new("server")
#initServer(s)

}
