% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.server.R
\name{print.server}
\alias{print.server}
\title{S3 method to print server objects (S4)}
\usage{
\method{print}{server}(x, ...)
}
\arguments{
\item{x}{server object}

\item{...}{parameters of the print generic method}
}
\value{
a string of character, a URL
}
\description{
A print method for \emph{serverURL} objects
}
\examples{

s <- new("server")
print(s)

}
