% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{push}
\alias{push}
\title{Pushes new view.json or data.json to the visualization server}
\usage{
push(visualization, type, list)
}
\arguments{
\item{visualization}{the visualization to be updated}

\item{type}{the type of file to be update, view or data}

\item{list}{the list object that contains the view or data}
}
\value{
an updated visualization object and new data/view.json on the server
}
\description{
Pushes new view.json or data.json to the visualization server
}
\examples{

x <- seq(from = 0, to = pi, by = 0.1)

#simple array
d <- appendData(variableName = "simpleArray", variable = cos(x), type = "1D Y array")
d <- appendData(data = d, variableName = "simpleArray2", variable = sin(x), type = "1D Y array")
v <- new("visualization")
push(v, type="data", d)

}
\seealso{
\pkg{visualizeR::saveJSON()}
}
