% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readView.R
\name{readView}
\alias{readView}
\title{Read json objects from https://github.com/npellet/visualizer.}
\usage{
readView(pathToFile)
}
\arguments{
\item{pathToFile}{the path to file}
}
\value{
a list object to be manipulated
}
\description{
Read json objects from https://github.com/npellet/visualizer.
}
\examples{
path <- file.path( system.file(package = "hastaLaVista"), "visu", "view", "blank.view.json" )
v <- readView( path )
}
