% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repCol.R
\name{repCol}
\alias{repCol}
\title{repeat a column when needed to match other matrix dimension}
\usage{
repCol(c, n)
}
\arguments{
\item{c}{a vector to be repeated}

\item{n}{number ot time to repeat the vector}
}
\value{
a matrix
}
\description{
repeat a column when needed to match other matrix dimension
}
\examples{

repCol(c(1,2,3), 2)
}
\references{
the original function was found here: \url{https://www.r-bloggers.com/a-quick-way-to-do-row-repeat-and-col-repeat-rep-row-rep-col/}
}
