% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassServer.R
\docType{class}
\name{server-class}
\alias{server-class}
\title{An S4 class for server}
\value{
an url object
}
\description{
An S4 class for server
}
\section{Slots}{

\describe{
\item{\code{baseURL}}{the url or IP of the server}

\item{\code{port}}{the port to which the server listen}

\item{\code{path}}{the path relative to the server root directory}

\item{\code{protocole}}{a protocole}

\item{\code{rootDir}}{root directory where to install the visualizer application}

\item{\code{init}}{if TRUE an instance of \pkg{servr::httw()} will be started to serve the files}
}}

\examples{

new("server")

}
