% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startWSServer.r
\name{startWSServer}
\alias{startWSServer}
\title{start a websocket server}
\usage{
startWSServer(port)
}
\arguments{
\item{port}{port if not default}
}
\value{
a connection
}
\description{
This function established a two way connection between R and the visualizer
using a websocket
}
\examples{
# conn <- startWSServer()
}
