% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useThisFolder.R
\name{useThisFolder}
\alias{useThisFolder}
\title{S4 method to use local folder to serve view and data for visualization objects (S4)}
\usage{
useThisFolder(v)
}
\arguments{
\item{v}{visualization object}
}
\value{
void
}
\description{
A method that add the current folder to the server path. .
}
\examples{

v<- new("visualization")
#useThisFolder(v)

}
