% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassVisualization.R
\docType{class}
\name{visualization-class}
\alias{visualization-class}
\title{An S4 class that contains a visualization}
\value{
a visualization object
}
\description{
An S4 class that contains a visualization
}
\section{Slots}{

\describe{
\item{\code{visuServer}}{server object that defines how to serve the \emph{github.com/npellet/visualizer} tool}

\item{\code{viewServer}}{server object that defines how to serve local view.json}

\item{\code{dataServer}}{server object that defines how to serve local data.json}

\item{\code{data}}{name of data json file}

\item{\code{view}}{name of view json file}
}}

\examples{

new("visualization")

}
