% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.visualization.R
\name{visualize,visualization-method}
\alias{visualize,visualization-method}
\title{S4 method to visualize visualization objects}
\usage{
\S4method{visualize}{visualization}(v, ...)
}
\arguments{
\item{v}{visualization object}

\item{...}{allows to pass translate = TRUE for running into docker instance}
}
\value{
void
}
\description{
A visualize method for \emph{visualization} objects. If \emph{object@localServer} is TRUE, then this
function will trigger a local instance of \pkg{servr} server.
}
\examples{

v<- new("visualization")
#visualize(v)

}
